requirements.txt
absl-py==0.8.0
aniso8601==8.0.0
astor==0.8.0
attrs==19.3.0
boto3==1.12.14
botocore==1.15.14
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cheroot==6.6.0
Click==7.0
comet-git-pure==0.19.14
comet-ml==3.0.2
configobj==5.0.6
cryptography==2.8
cycler==0.10.0
# cloudpickle
decorator==4.4.1
dgl==0.4.3
docutils==0.15.2
everett==1.0.2
filelock==3.0.12
Flask==1.1.1
flask-restplus==0.13.0
future==0.18.2
gast==0.2.2
ghast==0.1.0
google-pasta==0.1.7
grpcio==1.24.0
# gym==0.15.3
h5py==2.10.0
idna==2.9
imageio==2.6.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib
more-itertools==7.2.0
netifaces==0.10.9
networkx==2.4
nltk==3.4.5
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencv-python
opt-einsum==3.1.0
periodictable==1.5.1
Pillow==6.2.1
protobuf==3.9.2
pycparser==2.20
pyglet==1.3.2
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.2.post1
scipy==1.3.1
sentencepiece==0.1.86
six==1.14.0
sklearn==0.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-probability==0.8.0
termcolor==1.1.0
tokenizers==0.0.11
tqdm==4.43.0
transformers==2.4.1
urllib3==1.25.7
websocket-client==0.57.0
Werkzeug==0.16.0
# wrapt==1.12.0
# wurlitzer==2.0.0
# zipp
